Inherit from TTTableSubtitleItem/Cell.
The only problem with TTTableSubtitleItem/Cell is that it renders the
image/text/subtitle in a way we don't particularly want.  But the fields
it uses are fine, so we might as well inherit from it, and just override
the layout code, instead of replicating the field definitions too.

Also properly compute the table cell height.