The JSONValue extension gets angry with new Three20 versions.
Categories (such as the JSONValue method for NSString) need to be linked
with -ObjC, but new versions of Three20 get mad if they're linked with
that flag.  Or maybe the problem is related to having Three20 linked to
libjson.a rather than linking the Cineti executable to it directly.

Rather than figure that problem out, let's just not use the
stupid category method.