Hide navigation bar in movie list.
It takes up valuable space, we don't need it with the tab bar anyway,
and it also confuses TTLauncher slightly, leading to cut off posters and
a lack of pager dots.  Also adjust the background away from pure white
so that the active pager dot is visible, and animate the transition to
the movie view (just a slide for now, may get more elaborate later).