Fix animation of hiding/showing navigation bar.
If we're switching from the theater list tab to the movie list tab, we
don't want to animate the hiding of the navigation bar - it looks funny.
We *do* want to animate it when switching from the movie list to the
individual movie page and back again - otherwise the space where it
would be is black in the slide animation, then fills in later, which
looks gross.

Anyway, simple solution, just respect the animated parameter to
viewWillAppear: / viewWillDisappear: